Relocate CLI filesystem artifact helper into shared utilities
Move the writeFileArtifact/writeJsonArtifact helper from src/cli/src/modules/fs-artifacts.js into src/cli/src/shared/fs-artifacts.js so shared utilities no longer re-export from a module-local implementation. Update the manual and memory modules to import the shared helper in its new home.
