+++ src/cli/src/modules/manual/file-helpers.js
@@ -1,4 +1,4 @@
 export {
     writeFileArtifact as writeManualFile,
     writeJsonArtifact as writeManualJsonArtifact
-} from "../fs-artifacts.js";
+} from "../../shared/fs-artifacts.js";
